requirements.txt
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
mysql-connector-python==8.4.0
PyYAML==6.0.1
yarl==1.9.4
